feat(identity): keep the current global JWT signing/verification logic..
...and instead introduce another layer to handle both global and
identity provider specific verification (for JWKS)
