+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/client/TransportClient.java
@@ -33,7 +33,6 @@ import com.b2international.snowowl.core.api.SnowowlRuntimeException;
 import com.b2international.snowowl.core.api.SnowowlServiceException;
 import com.b2international.snowowl.core.authorization.AuthorizedEventBus;
 import com.b2international.snowowl.core.config.SnowOwlConfiguration;
-import com.b2international.snowowl.core.identity.AuthorizationHeaderVerifier;
 import com.b2international.snowowl.core.identity.User;
 import com.b2international.snowowl.core.identity.request.UserRequests;
 import com.b2international.snowowl.core.setup.Environment;
@@ -225,18 +224,17 @@ public final class TransportClient implements IDisposableService {
 			initConnection();
 			
 			// try to log in with the specified username and password using the non-authorized bus instance
-			final String accessToken = UserRequests.prepareLogin()
+			final User user = UserRequests.prepareLogin()
 				.setUsername(username)
 				.setPassword(password)
 				.buildAsync()
 				.execute(bus)
-				.getSync(3L, TimeUnit.SECONDS)
-				.getAccessToken();
+				.getSync(3L, TimeUnit.SECONDS);
 			
 			// if successfully logged in replace the event bus with an authorized one
-			env.services().registerService(IEventBus.class, new AuthorizedEventBus(bus, ImmutableMap.of("Authorization", accessToken)));
+			env.services().registerService(IEventBus.class, new AuthorizedEventBus(bus, ImmutableMap.of("Authorization", user.getAccessToken())));
 			
-			return env.service(AuthorizationHeaderVerifier.class).toUser(accessToken);
+			return user;
 		} catch (UnauthorizedException e) {
 			throw new SnowowlServiceException(e.getMessage());
 		} catch (final Throwable t) {
