--- core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/identity/JWTCapableIdentityProviderConfig.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2022 B2i Healthcare Pte Ltd, http://b2i.sg
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.b2international.snowowl.core.identity;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-/**
- * @since 8.8.0
- */
-public abstract class JWTCapableIdentityProviderConfig implements IdentityProviderConfig {
-
-	private JWTConfiguration jwt;
-	
-	@JsonProperty("jwt")
-	public JWTConfiguration getJWTConfiguration() {
-		return jwt;
-	}
-	
-	@JsonProperty("jwt")
-	public void setJWTConfiguration(JWTConfiguration jwt) {
-		this.jwt = jwt;
-	}
-	
-}
