+++ core/com.b2international.snowowl.identity.ldap/src/com/b2international/snowowl/identity/ldap/LdapIdentityProvider.java
@@ -32,7 +32,7 @@ import org.slf4j.LoggerFactory;
 
 import com.b2international.snowowl.core.api.SnowowlRuntimeException;
 import com.b2international.snowowl.core.events.util.Promise;
-import com.b2international.snowowl.core.identity.JWTCapableIdentityProvider;
+import com.b2international.snowowl.core.identity.IdentityProvider;
 import com.b2international.snowowl.core.identity.Permission;
 import com.b2international.snowowl.core.identity.User;
 import com.b2international.snowowl.core.identity.Users;
@@ -62,7 +62,7 @@ import com.google.common.collect.Iterators;
  * @since 5.11
  * @see LdapIdentityProviderConfig
  */
-final class LdapIdentityProvider extends JWTCapableIdentityProvider<LdapIdentityProviderConfig> {
+final class LdapIdentityProvider implements IdentityProvider {
 
 	private enum EmptyNamingEnumeration implements NamingEnumeration<Object> {
 		INSTANCE;
@@ -102,8 +102,10 @@ final class LdapIdentityProvider extends JWTCapableIdentityProvider<LdapIdentity
 	private static final String ATTRIBUTE_DN = "dn";
 	private static final String ATTR_CN = "cn";
 	
+	private final LdapIdentityProviderConfig conf;
+	
 	public LdapIdentityProvider(LdapIdentityProviderConfig conf) {
-		super(conf);
+		this.conf = conf;
 		final Map<String, String> options = new TreeMap<>();
 		options.put("bindDn", conf.getBindDn());
 		options.put("baseDn", conf.getBaseDn());
@@ -128,12 +130,7 @@ final class LdapIdentityProvider extends JWTCapableIdentityProvider<LdapIdentity
 	
 	@Override
 	public String getInfo() {
-		return String.join("@", TYPE, getConfiguration().getUri());
-	}
-	
-	@Override
-	protected String getType() {
-		return TYPE;
+		return String.join("@", TYPE, conf.getUri());
 	}
 	
 	@Override
@@ -158,7 +155,6 @@ final class LdapIdentityProvider extends JWTCapableIdentityProvider<LdapIdentity
 		Preconditions.checkNotNull(context, "Directory context is null.");
 		Preconditions.checkNotNull(username, "Username is null.");
 
-		final LdapIdentityProviderConfig conf = getConfiguration();
 		final String userFilterWithUsername = String.format("(&%s(%s=%s))", conf.getUserFilter(), conf.getUserIdProperty(), username);
 
 		NamingEnumeration<SearchResult> searchResultEnumeration = null;
@@ -194,7 +190,6 @@ final class LdapIdentityProvider extends JWTCapableIdentityProvider<LdapIdentity
 
 	@Override
 	public Promise<Users> searchUsers(Collection<String> usernames, int limit) {
-		final LdapIdentityProviderConfig conf = getConfiguration();
 		final ImmutableList.Builder<User> resultBuilder = ImmutableList.builder();
 		final String uidProp = conf.getUserIdProperty();
 		
@@ -238,7 +233,6 @@ final class LdapIdentityProvider extends JWTCapableIdentityProvider<LdapIdentity
 	}
 	
 	protected Collection<LdapRole> getAllLdapRoles(InitialLdapContext context) throws NamingException {
-		final LdapIdentityProviderConfig conf = getConfiguration();
 		NamingEnumeration<SearchResult> enumeration = null;
 		try {
 			final ImmutableList.Builder<LdapRole> results = ImmutableList.builder();
@@ -293,7 +287,6 @@ final class LdapIdentityProvider extends JWTCapableIdentityProvider<LdapIdentity
 	 * @return a Hashtable of properties in a format required by InitialLdapContext's constructor
 	 */
 	private Hashtable<String, Object> createLdapEnvironment() {
-		final LdapIdentityProviderConfig conf = getConfiguration();
 		final Hashtable<String, Object> env = new Hashtable<String, Object>();
 
 		env.put(Context.INITIAL_CONTEXT_FACTORY, LDAP_CTX_FACTORY);
